🧹 Major project cleanup and organization
📁 Structure Changes:
- Move all documentation from theme to root /docs/
- Organize gallery import scripts to /scripts/gallery-import/
- Remove unnecessary temporary and debug files
- Clean up root directory structure

📚 Documentation:
- Comprehensive new README.md with project overview
- Created docs/README.md as documentation index
- Added scripts/gallery-import/README.md for import guides
- Better organization and navigation

🗑️ Cleanup:
- Removed debug files (debug-gallery-admin.php, fix-gallery-permalink.php)
- Removed backup files and temporary scripts
- Removed duplicate documentation files
- Enhanced .gitignore for better file management

✨ Benefits:
- Cleaner project structure
- Better documentation organization
- Easier navigation for developers and users
- Improved maintainability
