+++ docs/PROTOTYPE_MIGRATION_ANALYSIS.md
@@ -1,0 +1,165 @@
+# 🗂️ ANALISIS DAMPAK PENGHAPUSAN FOLDER PROTOTYPE
+
+## ⚠️ DAMPAK MENGHAPUS FOLDER PROTOTYPE:
+
+### **1. BROKEN FUNCTIONALITY (HIGH IMPACT)**
+
+#### JavaScript Dependencies:
+- ❌ **functions.php line 194**: `prototype/js/index.js` - Header scroll, mobile menu toggle
+- ❌ **functions.php line 196**: `prototype/js/countdown.js` - Countdown timer functionality
+
+#### Image Assets:
+- ❌ **header.php line 102**: `prototype/images/logo.png` - Logo MTQ utama
+- ❌ **front-page.php line 221**: `prototype/images/gub-wagub-aceh.png` - Foto Gubernur
+- ❌ **front-page.php line 256**: `prototype/images/bupati-dan-wakil-2025.png` - Foto Bupati
+
+#### Build Configuration:
+- ❌ **tailwind.config.js line 5-6**: Tailwind scanning prototype files
+
+### **2. FITUR YANG AKAN RUSAK:**
+
+1. **Header Navigation**:
+   - Mobile menu tidak berfungsi
+   - Header scroll effects hilang
+   - Logo MTQ tidak tampil
+
+2. **Countdown Timer**:
+   - Countdown acara MTQ tidak berfungsi
+   - Error JavaScript di halaman frontend
+
+3. **Leadership Section**:
+   - Foto Gubernur dan Bupati tidak tampil
+   - Layout section rusak
+
+4. **Build Process**:
+   - Tailwind CSS mungkin tidak build dengan benar
+
+## ✅ SOLUSI & RENCANA MIGRASI:
+
+### **STEP 1: Migrasi Assets**
+
+#### A. Migrasi JavaScript Files:
+```bash
+# Copy JavaScript files ke folder assets
+cp wp-content/themes/mtq-aceh-pidie-jaya/prototype/js/index.js wp-content/themes/mtq-aceh-pidie-jaya/assets/js/
+cp wp-content/themes/mtq-aceh-pidie-jaya/prototype/js/countdown.js wp-content/themes/mtq-aceh-pidie-jaya/assets/js/
+```
+
+#### B. Images Already Migrated ✅:
+- Logo: `assets/images/logo.png` ✅
+- Gubernur: `assets/images/gub-wagub-aceh.png` ✅  
+- Bupati: `assets/images/bupati-dan-wakil-2025.png` ✅
+
+### **STEP 2: Update References**
+
+#### A. Update functions.php:
+```php
+// OLD (prototype):
+wp_enqueue_script('mtq-aceh-pidie-jaya-prototype-js', get_template_directory_uri() . '/prototype/js/index.js', array(), _S_VERSION, true);
+wp_enqueue_script('mtq-aceh-pidie-jaya-countdown', get_template_directory_uri() . '/prototype/js/countdown.js', array(), _S_VERSION, true);
+
+// NEW (assets):
+wp_enqueue_script('mtq-aceh-pidie-jaya-main-js', get_template_directory_uri() . '/assets/js/index.js', array(), _S_VERSION, true);
+wp_enqueue_script('mtq-aceh-pidie-jaya-countdown', get_template_directory_uri() . '/assets/js/countdown.js', array(), _S_VERSION, true);
+```
+
+#### B. Update Template Files:
+```php
+// header.php - Change:
+src="<?php echo get_template_directory_uri(); ?>/prototype/images/logo.png"
+// To:
+src="<?php echo get_template_directory_uri(); ?>/assets/images/logo.png"
+
+// front-page.php - Change:
+src="<?php echo get_template_directory_uri(); ?>/prototype/images/gub-wagub-aceh.png"
+src="<?php echo get_template_directory_uri(); ?>/prototype/images/bupati-dan-wakil-2025.png"
+// To:
+src="<?php echo get_template_directory_uri(); ?>/assets/images/gub-wagub-aceh.png"
+src="<?php echo get_template_directory_uri(); ?>/assets/images/bupati-dan-wakil-2025.png"
+```
+
+#### C. Update tailwind.config.js:
+```javascript
+// Remove prototype references:
+"./wp-content/themes/mtq-aceh-pidie-jaya/prototype/**/*.html",
+"./wp-content/themes/mtq-aceh-pidie-jaya/prototype/js/**/*.js"
+
+// Keep only active files:
+"./wp-content/themes/mtq-aceh-pidie-jaya/**/*.php",
+"./wp-content/themes/mtq-aceh-pidie-jaya/assets/js/**/*.js"
+```
+
+### **STEP 3: Testing Before Deletion**
+
+#### Test Checklist:
+- [ ] Header logo tampil
+- [ ] Mobile menu berfungsi
+- [ ] Header scroll effect bekerja
+- [ ] Countdown timer running
+- [ ] Foto Gubernur/Bupati tampil
+- [ ] Tailwind build sukses
+- [ ] No JavaScript errors
+
+## 🎯 RECOMMENDED ACTION PLAN:
+
+### **Phase 1: Preparation (SAFE)**
+1. Copy JavaScript files to assets/js/
+2. Update all template references
+3. Update functions.php enqueue
+4. Update tailwind.config.js
+5. Test functionality
+
+### **Phase 2: Validation (CRITICAL)**
+1. Run `npm run build` - check for errors
+2. Test frontend functionality
+3. Check browser console for errors
+4. Verify all images load
+5. Test mobile navigation
+
+### **Phase 3: Safe Deletion (FINAL)**
+1. Create backup: `tar -czf prototype-backup.tar.gz prototype/`
+2. Delete folder: `rm -rf prototype/`
+3. Final testing
+4. Monitor for issues
+
+## 📋 MIGRATION SCRIPT:
+
+```bash
+#!/bin/bash
+# MTQ Prototype Migration Script
+
+THEME_DIR="wp-content/themes/mtq-aceh-pidie-jaya"
+
+echo "🚀 Starting prototype migration..."
+
+# 1. Copy JavaScript files
+echo "📁 Copying JavaScript files..."
+cp $THEME_DIR/prototype/js/index.js $THEME_DIR/assets/js/
+cp $THEME_DIR/prototype/js/countdown.js $THEME_DIR/assets/js/
+
+# 2. Backup prototype folder
+echo "💾 Creating backup..."
+tar -czf prototype-backup-$(date +%Y%m%d).tar.gz $THEME_DIR/prototype/
+
+echo "✅ Migration preparation complete!"
+echo "📝 Next: Update code references manually"
+```
+
+## ⚠️ CRITICAL NOTES:
+
+1. **DO NOT DELETE** prototype folder until migration complete
+2. **TEST THOROUGHLY** before final deletion
+3. **BACKUP FIRST** - use the provided script
+4. **Images already migrated** to assets/images ✅
+5. **JavaScript needs manual migration** to assets/js
+
+## 🎯 IMPACT SUMMARY:
+
+- **HIGH RISK** if deleted without migration
+- **ZERO RISK** after proper migration
+- **All assets have alternatives** in assets folder
+- **Clean migration path available**
+
+---
+*Analysis Date: September 13, 2025*  
+*Status: Ready for migration*
