+++ docs/README.md
@@ -1,0 +1,59 @@
+# 📖 Project Documentation Index
+
+Selamat datang di dokumentasi lengkap proyek MTQ Aceh Pidie Jaya 2025.
+
+## 📚 **Dokumentasi Utama**
+
+### **User Guides**
+- [**Gallery System Complete Guide**](./GALLERY_SYSTEM_DOCUMENTATION.md) - Panduan lengkap penggunaan sistem gallery
+- [**Gallery Admin Panel Fix**](./GALLERY_ADMIN_PANEL_FIX.md) - Troubleshooting admin panel gallery
+- [**Gallery Fix Solution**](./GALLERY_FIX_SOLUTION.md) - Solusi masalah umum gallery
+
+### **Development Guides**
+- [**Theme Development Plan**](../THEME_DEVELOPMENT_PLAN.md) - Rencana pengembangan theme
+- [**Deployment Guide**](../DEPLOYMENT_GUIDE.md) - Panduan deployment ke production
+- [**Import Guide**](../IMPORT_GUIDE.md) - Cara import data gallery
+
+### **Technical Documentation**
+- [**Changelog**](./CHANGELOG.md) - Riwayat perubahan dan update
+- [**Social Sharing README**](./SOCIAL_SHARING_README.md) - Dokumentasi fitur social sharing
+- [**Countdown Configuration**](./README-COUNTDOWN-CONFIG.md) - Setup countdown timer
+
+### **Fix & Troubleshooting**
+- [**Live Server Gallery Fix**](./LIVE_SERVER_GALLERY_FIX.md) - Perbaikan masalah server live
+- [**Custom Header Fix**](./CUSTOM_HEADER_FIX_SUMMARY.md) - Perbaikan header custom
+- [**YouTube Live Status**](./YOUTUBE_LIVE_STATUS.md) - Status dan troubleshooting YouTube live
+
+### **Archive Documentation**
+- [**Git Commit Summary**](./GIT_COMMIT_SUMMARY.md) - Summary commit penting
+- [**Prototype Migration**](./PROTOTYPE_MIGRATION_ANALYSIS.md) - Analisis migrasi prototype
+- [**YouTube Live Duplicate Fix**](./YOUTUBE_LIVE_DUPLICATE_FIX.md) - Perbaikan duplikasi YouTube live
+
+## 🚀 **Quick Links**
+
+### **Untuk Pengguna Baru:**
+1. Mulai dengan [Gallery System Documentation](./GALLERY_SYSTEM_DOCUMENTATION.md)
+2. Baca [Import Guide](../IMPORT_GUIDE.md) untuk import data
+3. Lihat [Deployment Guide](../DEPLOYMENT_GUIDE.md) untuk setup
+
+### **Untuk Developer:**
+1. Baca [Theme Development Plan](../THEME_DEVELOPMENT_PLAN.md)
+2. Check [Changelog](./CHANGELOG.md) untuk update terbaru
+3. Lihat [Technical Guides](#technical-documentation) untuk detail teknis
+
+### **Untuk Troubleshooting:**
+1. [Gallery Issues](./GALLERY_FIX_SOLUTION.md) - Masalah gallery umum
+2. [Admin Panel Issues](./GALLERY_ADMIN_PANEL_FIX.md) - Masalah admin panel
+3. [Live Server Issues](./LIVE_SERVER_GALLERY_FIX.md) - Masalah server production
+
+## 📞 **Need Help?**
+
+Jika dokumentasi di atas tidak menjawab pertanyaan Anda:
+
+1. **Check GitHub Issues:** [Project Issues](https://github.com/saipulbahri-it/wp-mtq-aceh-pidie-jaya/issues)
+2. **Create New Issue:** Dengan label yang sesuai
+3. **Contact Developer:** Saipul Bahri IT
+
+---
+
+*Dokumentasi ini diupdate secara berkala. Terakhir update: September 2025*
