--- mtq.pidiejayakab.go.id
@@ -1,85 +1,0 @@
-server {
-    listen 80;
-    listen [::]:80;
-    
-    server_name mtq.pidiejayakab.go.id;
-    
-    # Redirect all HTTP requests to HTTPS
-    return 301 https://$server_name$request_uri;
-}
-
-server {
-    listen 443 ssl;
-    listen [::]:443 ssl;
-    http2 on;
-    
-    server_name mtq.pidiejayakab.go.id;
-    
-    # SSL certificate settings
-    ssl_certificate /opt/homebrew/etc/nginx/ssl/mtq.pidiejayakab.go.id.crt;
-    ssl_certificate_key /opt/homebrew/etc/nginx/ssl/mtq.pidiejayakab.go.id.key;
-    
-    # Security headers (note: removed HSTS for testing)
-    add_header X-Frame-Options "SAMEORIGIN" always;
-    add_header X-XSS-Protection "1; mode=block" always;
-    add_header X-Content-Type-Options "nosniff" always;
-    # add_header Strict-Transport-Security "max-age=31536000; includeSubDomains" always;
-    
-    # Root directory and index files
-    root /Users/saipulbahri.it/Projects/github/web-mtq-pijay/wp-mtq-aceh-pidie-jaya;
-    index index.php index.html index.htm;
-    
-    # WordPress specific rules
-    location / {
-        try_files $uri $uri/ /index.php$is_args$args;
-    }
-    
-    # Hide sensitive files
-    location ~ /\. {
-        deny all;
-    }
-    
-    location ~ ^/wp-config\.php {
-        deny all;
-    }
-    
-    location ~ ^/wp-content/uploads/.*\.php$ {
-        deny all;
-    }
-    
-    # Handle PHP files
-    location ~ \.php$ {
-        try_files $uri =404;
-        fastcgi_split_path_info ^(.+\.php)(/.+)$;
-        fastcgi_pass 127.0.0.1:9083;  # PHP 8.3 FPM port
-        fastcgi_index index.php;
-        include fastcgi_params;
-        fastcgi_param SCRIPT_FILENAME $document_root$fastcgi_script_name;
-        fastcgi_param HTTP_PROXY "";  # Mitigate CVE-2016-5385, CVE-2016-5387
-        fastcgi_intercept_errors off;
-        fastcgi_buffer_size 128k;
-        fastcgi_buffers 4 256k;
-        fastcgi_busy_buffers_size 256k;
-    }
-    
-    # Deny access to htaccess files
-    location ~ /\.ht {
-        deny all;
-    }
-    
-    # Cache static files
-    location ~* \.(css|js|png|jpg|jpeg|gif|ico|svg|woff|woff2|ttf|eot)$ {
-        expires 1y;
-        add_header Cache-Control "public, immutable";
-        log_not_found off;
-    }
-    
-    # Security: deny access to readme.html, license.txt and other sensitive files
-    location ~* (readme\.html|license\.txt|wp-config\.php) {
-        deny all;
-    }
-    
-    # Logging
-    access_log /opt/homebrew/var/log/nginx/mtq.pidiejayakab.go.id.access.log;
-    error_log /opt/homebrew/var/log/nginx/mtq.pidiejayakab.go.id.error.log;
-}
