--- tasks.md
@@ -1,145 +1,0 @@
-## 📋 **TASKS LIST - TEMA MTQ ACEH PIDIE JAYA**
-
-### **🔥 HIGH PRIORITY TASKS**
-
-#### **1. Template Files yang Hilang**
-- [ ] **Buat archive.php** - Template untuk archive kategori, tags, dan author
-- [x] **Buat category.php** - Template khusus untuk halaman kategori  
-- [ ] **Buat tag.php** - Template khusus untuk halaman tags
-- [ ] **Buat search.php** - Template untuk hasil pencarian
-- [ ] **Buat searchform.php** - Form pencarian custom
-- [ ] **Buat author.php** - Template untuk archive penulis
-- [ ] **Buat date.php** - Template untuk archive berdasarkan tanggal
-
-#### **2. Bug Fixes Critical**
-- [ ] **Test mobile navigation** - Pastikan menu mobile berfungsi dengan baik
-- [ ] **Validate template-tags functions** - Pastikan semua fungsi template bekerja
-
-#### **3. Essential Functionality**
-- [ ] **Implementasi breadcrumbs** - Navigation breadcrumb di seluruh template
-- [ ] **Related posts system** - Tambahkan related posts di single.php
-- [ ] **Search widget optimization** - Perbaiki search widget di 404.php
-
-### **⚡ MEDIUM PRIORITY TASKS**
-
-#### **4. User Experience Enhancements**
-- [ ] **Social sharing buttons** - Tambahkan tombol share di single posts
-- [ ] **Comment system styling** - Style form komentar dan tampilan comments
-- [ ] **Print stylesheet** - CSS khusus untuk print
-- [ ] **Loading states** - Improve loading indicators
-- [ ] **Error handling** - Better error messages dan fallbacks
-
-#### **5. Accessibility & Mobile**
-- [ ] **Keyboard navigation** - Focus states dan tab navigation
-- [ ] **ARIA labels** - Tambahkan ARIA attributes yang lengkap
-- [ ] **Skip links functionality** - Implementasi skip to content
-- [ ] **Screen reader optimization** - Text untuk screen readers
-- [ ] **Touch gestures** - Optimize untuk touch devices
-
-#### **6. Content Management**
-- [ ] **Post formats support** - Video, audio, gallery, quote formats
-- [ ] **Custom post types** - Jika diperlukan untuk MTQ events
-- [ ] **Archive pagination** - Implement proper pagination di archive pages
-- [ ] **Content filtering** - Filter posts by category/tags di archive
-
-### **🎨 LOW PRIORITY TASKS**
-
-#### **7. Advanced Customizer Options**
-- [ ] **Logo upload customizer** - Upload custom logo
-- [ ] **Color scheme options** - Customize theme colors
-- [ ] **Typography options** - Font family dan size options
-- [ ] **Social media links** - Customizer untuk social links
-- [ ] **Contact information** - Footer contact info customizer
-- [ ] **Header/Footer widgets** - Additional widget areas
-
-#### **8. Performance Optimization**
-- [ ] **Image lazy loading** - Implement lazy loading untuk images
-- [ ] **Critical CSS inline** - Inline critical CSS
-- [ ] **Icon fonts optimization** - Optimize atau ganti dengan SVG
-- [ ] **JS/CSS minification** - Minify untuk production
-- [ ] **Caching integration** - Support untuk caching plugins
-- [ ] **WebP image support** - Modern image format support
-
-#### **9. SEO & Analytics**
-- [ ] **Schema.org markup** - Rich snippets untuk better SEO
-- [ ] **Open Graph tags** - Social media sharing optimization
-- [ ] **Twitter Cards** - Twitter sharing optimization
-- [ ] **Analytics integration** - Google Analytics/Tag Manager support
-- [ ] **Sitemap support** - XML sitemap generation
-
-### **🔒 SECURITY & VALIDATION TASKS**
-
-#### **10. Security Enhancements**
-- [ ] **Input sanitization** - Validate semua user inputs
-- [ ] **Nonce verification** - Tambahkan nonces untuk forms
-- [ ] **Escape output** - Escape semua dynamic content
-- [ ] **SQL injection protection** - Secure database queries
-- [ ] **XSS protection** - Prevent cross-site scripting
-
-#### **11. Code Quality**
-- [ ] **PHP coding standards** - Follow WordPress coding standards
-- [ ] **CSS organization** - Organize dan comment CSS code
-- [ ] **JS optimization** - Clean dan optimize JavaScript
-- [ ] **Documentation** - Add inline documentation
-- [ ] **Error logging** - Implement proper error handling
-
-### **🌐 INTERNATIONALIZATION TASKS**
-
-#### **12. Translation Ready**
-- [ ] **Generate .pot file** - Translation template file
-- [ ] **RTL language support** - Right-to-left languages
-- [ ] **More text domains** - Cover semua text dengan text domain
-- [ ] **Translation functions** - Ensure all text is translatable
-- [ ] **Language packs** - Create Indonesian language pack
-
-### **🧪 TESTING & QA TASKS**
-
-#### **13. Testing & Quality Assurance**
-- [ ] **Cross-browser testing** - Test di berbagai browser
-- [ ] **Device testing** - Test di berbagai ukuran layar
-- [ ] **Performance testing** - Page speed dan loading times
-- [ ] **Accessibility testing** - WCAG compliance testing
-- [ ] **WordPress compatibility** - Test dengan berbagai versi WP
-- [ ] **Plugin compatibility** - Test dengan plugin populer
-
-### **📚 DOCUMENTATION TASKS**
-
-#### **14. Documentation**
-- [ ] **User manual** - Panduan penggunaan tema
-- [ ] **Developer documentation** - Technical documentation
-- [ ] **Customization guide** - Cara customize tema
-- [ ] **Troubleshooting guide** - Common issues dan solutions
-- [ ] **Changelog** - Version history dan updates
-
-### **🚀 DEPLOYMENT TASKS**
-
-#### **15. Production Ready**
-- [ ] **Version control** - Proper Git workflow
-- [ ] **Build process** - Automated build untuk production
-- [ ] **Asset optimization** - Compress images dan files
-- [ ] **Backup system** - Backup strategy
-- [ ] **Update mechanism** - Theme update system
-
----
-
-## 📊 **TASK PRIORITY MATRIX:**
-
-**🔴 CRITICAL (Do First):**
-- Fix content-page.php bug
-- Create archive.php
-- Create search.php & searchform.php
-
-**🟡 IMPORTANT (Do Next):**  
-- Breadcrumbs system
-- Related posts
-- Mobile navigation validation
-
-**🟢 ENHANCEMENT (Do Later):**
-- Customizer options
-- Performance optimization
-- Advanced SEO features
-
-**🔵 FUTURE (Nice to Have):**
-- Custom post types
-- Advanced analytics
-- Multi-language support
