--- wp-content/themes/mtq-aceh-pidie-jaya/CHANGELOG.md
@@ -1,171 +1,0 @@
-# Changelog - MTQ Aceh Pidie Jaya Template
-
-## [v2.1.0] - 2024-01-15
-
-### ✨ Features Added
-- **Progress Bar Controls**: Added show/hide functionality for countdown progress bars
-  - Toggle visibility controls in admin panel
-  - Dynamic CSS class generation for progress bar states
-  - Real-time preview updates with progress bar toggle
-  - Consistent UI with other show/hide controls
-
-### 🔒 Security Enhancements
-- **Comprehensive Security Audit**: Achieved 91/100 security score
-- **Rate Limiting System**: 
-  - IP-based rate limiting for AJAX endpoints
-  - Configurable limits (10/min users, 20/min admin)
-  - Automatic blocking for abuse attempts
-- **Enhanced Input Validation**:
-  - Improved sanitization for all user inputs
-  - Stronger validation rules for admin forms
-  - Better error handling and feedback
-- **Security Headers Implementation**:
-  - X-Frame-Options: SAMEORIGIN
-  - X-Content-Type-Options: nosniff
-  - X-XSS-Protection: 1; mode=block
-  - Referrer-Policy: strict-origin-when-cross-origin
-- **File Protection**:
-  - Added .htaccess protection for `/scripts/` directory
-  - Added .htaccess protection for `/data/` directory
-  - Blocked direct access to sensitive files
-- **Security Logging**:
-  - Comprehensive security event logging
-  - Failed authentication tracking
-  - Suspicious activity monitoring
-  - Rate limit breach logging
-
-### 🛠️ Technical Improvements
-- **Code Quality**:
-  - Enhanced error handling in countdown admin
-  - Better method organization and documentation
-  - Improved code comments and inline docs
-- **Performance**:
-  - Optimized AJAX request handling
-  - Better caching for transient data
-  - Reduced database queries for rate limiting
-
-### 📚 Documentation
-- **SECURITY_AUDIT_REPORT.md**: Comprehensive security analysis
-- **SECURITY_GUIDE.md**: Developer security guidelines
-- **Updated code comments**: Better inline documentation
-
-### 🐛 Bug Fixes
-- Fixed potential XSS vulnerabilities in admin preview
-- Resolved CSRF protection issues in AJAX handlers
-- Fixed function duplication in security enhancements
-- Improved error handling for invalid inputs
-
-### 🔧 Configuration Changes
-- Enhanced nonce verification for all AJAX actions
-- Improved capability checking for admin functions
-- Better sanitization patterns throughout codebase
-
----
-
-## [v2.0.0] - 2024-01-14
-
-### ✨ Major Features
-- **Countdown System**: Complete countdown timer implementation
-  - Real-time countdown display
-  - Admin configuration panel
-  - Custom styling and animations
-  - Progress indicators
-
-### 🎨 UI/UX Improvements
-- **Responsive Design**: Mobile-first countdown interface
-- **Admin Panel**: User-friendly configuration interface
-- **Visual Enhancements**: Modern card-based design
-- **Animations**: Smooth transitions and hover effects
-
-### 🛠️ Technical Foundation
-- **Admin Class Structure**: Object-oriented admin panel
-- **AJAX Integration**: Real-time preview updates
-- **CSS Framework**: Organized stylesheet structure
-- **WordPress Integration**: Proper hooks and filters
-
----
-
-## [v1.0.0] - 2024-01-01
-
-### 🎉 Initial Release
-- **Base Template**: WordPress theme foundation
-- **MTQ Branding**: Custom styling for MTQ Aceh Pidie Jaya
-- **Basic Functionality**: Core template features
-
----
-
-## Security Audit History
-
-### 2024-01-15: Comprehensive Security Review
-- **Audit Score**: 91/100
-- **Vulnerabilities Found**: 0 Critical, 2 Medium, 3 Low
-- **Fixes Applied**: All medium and high priority issues resolved
-- **Recommendations**: Additional security headers and logging implemented
-
-### Security Metrics:
-- ✅ Input Sanitization: 100%
-- ✅ Output Escaping: 98%
-- ✅ CSRF Protection: 100%
-- ✅ Access Control: 95%
-- ✅ File Security: 90%
-- ✅ SQL Injection Prevention: 100%
-- ✅ XSS Prevention: 98%
-
----
-
-## Development Notes
-
-### Code Quality Standards
-- WordPress Coding Standards compliance
-- PHPCS validation passed
-- Security best practices followed
-- Comprehensive error handling
-
-### Testing Coverage
-- ✅ Admin panel functionality
-- ✅ AJAX endpoint security
-- ✅ Rate limiting effectiveness
-- ✅ Cross-browser compatibility
-- ✅ Mobile responsiveness
-
-### Performance Metrics
-- Page Load Time: < 2s
-- AJAX Response Time: < 500ms
-- Security Check Overhead: < 50ms
-- Memory Usage: Optimized
-
----
-
-## Upgrade Instructions
-
-### From v2.0.0 to v2.1.0:
-1. Backup current theme files
-2. Update theme files
-3. Clear any caching plugins
-4. Test countdown functionality
-5. Verify security features are active
-
-### Security Upgrade Notes:
-- Rate limiting is automatically active
-- Security headers are applied globally
-- .htaccess files are auto-generated
-- No manual configuration required
-
----
-
-## Support & Maintenance
-
-### Regular Updates:
-- Security patches: Monthly
-- Feature updates: Quarterly
-- WordPress compatibility: With WP releases
-
-### Monitoring:
-- Security logs: Daily review
-- Performance metrics: Weekly analysis
-- User feedback: Continuous integration
-
----
-
-*Maintained by: MTQ Aceh Pidie Jaya Development Team*
-*Next scheduled review: 2024-02-15*
