--- wp-content/themes/mtq-aceh-pidie-jaya/CUSTOM_HEADER_ANALYSIS.md
@@ -1,213 +1,0 @@
-# 🔍 ANALISIS CUSTOM-HEADER.PHP
-
-## 📊 STATUS FILE:
-
-### **📍 Lokasi:**
-- `wp-content/themes/mtq-aceh-pidie-jaya/inc/custom-header.php`
-- **Size:** 96 lines
-- **Loaded in:** `functions.php` line 289
-
-## 🔍 ANALISIS KONTEN:
-
-### **1. FUNGSI UTAMA:**
-
-#### A. `mtq_aceh_pidie_jaya_custom_header_setup()`
-```php
-function mtq_aceh_pidie_jaya_custom_header_setup() {
-    add_theme_support('custom-header', array(
-        'default-image' => '',
-        'default-text-color' => '000000',
-        'width' => 1000,
-        'height' => 250,
-        'flex-height' => true,
-        'wp-head-callback' => 'mtq_aceh_pidie_jaya_header_style',
-    ));
-}
-```
-
-**Konfigurasi:**
-- ✅ **Width:** 1000px
-- ✅ **Height:** 250px (flexible)
-- ✅ **Default text color:** #000000 (black)
-- ✅ **Default image:** None
-- ✅ **Callback:** Custom styling function
-
-#### B. `mtq_aceh_pidie_jaya_header_style()`
-```php
-function mtq_aceh_pidie_jaya_header_style() {
-    // Handle custom header text color
-    // Generate inline CSS for header styling
-}
-```
-
-**Fungsi:**
-- 🎨 Handle custom header text colors
-- 📝 Generate inline CSS dynamically
-- 🔧 Support for hiding header text
-
-### **2. MASALAH YANG DITEMUKAN:**
-
-#### ❌ **Syntax Error di Line 70:**
-```php
-<?php echo '	<style type="text/css";>'; ?>
-                                  ^
-                              Semicolon salah tempat
-```
-
-**Problem:** Semicolon (`;`) seharusnya tidak ada di atribut HTML.
-**Should be:** `<style type="text/css">`
-
-#### ❌ **CSS Validation Filter Issue:**
-```php
-add_filter('wp_enqueue_scripts', function () {
-    if (is_admin()) {
-        return;
-    }
-    wp_add_inline_style('theme-style-handle', '/* CSS validation disabled */');
-}, 100);
-```
-
-**Problems:**
-1. **Wrong hook:** `wp_enqueue_scripts` is action, not filter
-2. **Invalid handle:** `theme-style-handle` not registered
-3. **Poor implementation:** Doesn't actually disable validation
-
-#### ⚠️ **Tidak Digunakan:**
-- Custom header functionality tidak diimplementasi di `header.php`
-- Tidak ada `the_header_image_tag()` call di template
-- Site title/description styling mungkin konflik dengan Tailwind
-
-### **3. IMPLEMENTASI SAAT INI:**
-
-#### A. WordPress Customizer:
-- ✅ **Available:** Admin dapat upload header image
-- ✅ **Text Color:** Can be customized
-- ❌ **Not Used:** Tidak tampil di frontend
-
-#### B. Frontend Integration:
-- ❌ **Missing:** Tidak ada di header.php
-- ❌ **Unused:** Custom header image tidak ditampilkan
-- ❌ **Conflict:** CSS styling mungkin bentrok dengan Tailwind
-
-## 🛠️ REKOMENDASI PERBAIKAN:
-
-### **1. FIX SYNTAX ERROR (CRITICAL):**
-
-#### Current (Line 70):
-```php
-<?php echo '	<style type="text/css";>'; ?>
-```
-
-#### Fixed:
-```php
-<style type="text/css">
-```
-
-### **2. FIX CSS VALIDATION FILTER:**
-
-#### Current (Lines 86-94):
-```php
-add_filter('wp_enqueue_scripts', function () {
-    if (is_admin()) {
-        return;
-    }
-    wp_add_inline_style('theme-style-handle', '/* CSS validation disabled */');
-}, 100);
-```
-
-#### Fixed:
-```php
-// Remove this section entirely or fix properly:
-add_action('wp_enqueue_scripts', function () {
-    if (is_admin()) {
-        return;
-    }
-    // Only if you have a registered style handle
-    // wp_add_inline_style('mtq-aceh-pidie-jaya-style', '/* Custom header styles */');
-}, 100);
-```
-
-### **3. INTEGRATE WITH HEADER.PHP:**
-
-#### Add to header.php (after opening header tag):
-```php
-<?php if (get_header_image()) : ?>
-    <div class="custom-header-image">
-        <?php the_header_image_tag(); ?>
-    </div>
-<?php endif; ?>
-```
-
-### **4. TAILWIND INTEGRATION:**
-
-#### Update CSS to work with Tailwind:
-```php
-function mtq_aceh_pidie_jaya_header_style() {
-    $header_text_color = get_header_textcolor();
-    
-    if (get_theme_support('custom-header', 'default-text-color') === $header_text_color) {
-        return;
-    }
-    
-    if (!display_header_text()) : ?>
-        <style type="text/css">
-        .site-title,
-        .site-description {
-            @apply sr-only;
-        }
-        </style>
-    <?php else : ?>
-        <style type="text/css">
-        .site-title a,
-        .site-description {
-            color: #<?php echo esc_attr($header_text_color); ?> !important;
-        }
-        </style>
-    <?php endif;
-}
-```
-
-## 🎯 PRIORITAS PERBAIKAN:
-
-### **HIGH PRIORITY:**
-1. ✅ **Fix syntax error** di line 70 (critical)
-2. ✅ **Remove/fix CSS validation filter** 
-3. ✅ **Clean up unused code**
-
-### **MEDIUM PRIORITY:**
-1. 🔧 **Integrate custom header** dengan frontend
-2. 🎨 **Add Tailwind-compatible styling**
-3. 📱 **Make responsive design**
-
-### **LOW PRIORITY:**
-1. 📚 **Add better documentation**
-2. 🧪 **Add customizer preview**
-3. 🎛️ **Enhanced customization options**
-
-## 📋 IMPACT ASSESSMENT:
-
-### **Jika Tidak Diperbaiki:**
-- ❌ **Syntax error** dapat menyebabkan PHP warnings
-- ❌ **Invalid CSS** di frontend
-- ❌ **Customizer confusion** - user bisa set header tapi tidak tampil
-- ❌ **Code quality** issues
-
-### **Setelah Diperbaiki:**
-- ✅ **Clean PHP code** tanpa syntax error
-- ✅ **Working custom header** functionality
-- ✅ **Better user experience** di customizer
-- ✅ **Consistent styling** dengan theme
-
-## 🎯 KESIMPULAN:
-
-**CUSTOM-HEADER.PHP PERLU DIPERBAIKI:**
-- Syntax error pada line 70 (critical)
-- CSS validation filter bermasalah
-- Functionality tidak terintegrasi dengan frontend
-- Potential conflicts dengan Tailwind CSS
-
-**RECOMMENDED ACTION:** Fix syntax errors dan integrate dengan header.php untuk functionality yang complete.
-
----
-*Analysis Date: September 13, 2025*  
-*Status: Needs immediate fixes*
