--- wp-content/themes/mtq-aceh-pidie-jaya/GALLERY_FIX_SOLUTION.md
@@ -1,86 +1,0 @@
-# 🔧 Solusi: Halaman Galeri Tidak Ditemukan
-
-## Masalah
-Ketika mengklik menu "Galeri" di header, muncul error **"Halaman Galeri Tidak Ditemukan"** atau **404 Not Found**.
-
-## Penyebab
-WordPress belum melakukan **flush permalink structure** setelah custom post type `mtq_gallery` didaftarkan. Ini adalah masalah umum saat menambah custom post type baru.
-
-## Solusi yang Telah Diterapkan
-
-### 1. ✅ **Perbaikan Otomatis di functions.php**
-```php
-// Initialize gallery system dengan proper permalink flush
-function mtq_init_gallery_system() {
-    new MTQ_Gallery_Post_Type();
-    new MTQ_Gallery_Shortcodes();
-    
-    // Flush permalinks jika belum dilakukan
-    if (get_option('mtq_gallery_permalinks_flushed') !== 'yes') {
-        flush_rewrite_rules();
-        update_option('mtq_gallery_permalinks_flushed', 'yes');
-    }
-}
-
-// Flush saat theme diaktifkan
-function mtq_theme_activation() {
-    delete_option('mtq_gallery_permalinks_flushed');
-    flush_rewrite_rules();
-}
-```
-
-### 2. ✅ **Admin Notice untuk Guidance**
-Ditambahkan notice di wp-admin yang memberikan panduan jika masih ada masalah.
-
-### 3. ✅ **Script Diagnostik**
-File `fix-gallery-permalink.php` untuk troubleshooting dan perbaikan manual.
-
-## Cara Mengatasi Jika Masih Bermasalah
-
-### Metode 1: **Manual Permalink Refresh** (Tercepat)
-1. Login ke **WP Admin**
-2. Pergi ke **Settings → Permalinks** 
-3. Klik **"Save Changes"** (tanpa mengubah apapun)
-4. Test kembali menu Galeri
-
-### Metode 2: **Menggunakan Script Diagnostik**
-1. Akses: `domain.com/wp-content/themes/mtq-aceh-pidie-jaya/fix-gallery-permalink.php`
-2. Script akan:
-   - Check apakah gallery post type terdaftar
-   - Flush permalink structure
-   - Test URL gallery
-   - Memberikan diagnostik lengkap
-
-### Metode 3: **Via WP-CLI** (untuk developer)
-```bash
-wp rewrite flush --hard
-wp post-type list
-```
-
-## Verifikasi Perbaikan
-
-### URLs yang Harus Berfungsi:
-- `domain.com/gallery/` - Halaman archive gallery
-- `domain.com/gallery/nama-gallery/` - Single gallery post
-
-### Template Files:
-- ✅ `archive-mtq_gallery.php` - Untuk halaman daftar gallery
-- ✅ `single-mtq_gallery.php` - Untuk single gallery post
-
-## Pencegahan Masalah Serupa
-
-1. **Selalu flush permalinks** setelah registrasi custom post type
-2. **Gunakan proper hooks** (`init`, `after_switch_theme`)
-3. **Test URL** setelah perubahan struktur permalink
-
-## Status Saat Ini
-
-✅ **Gallery system fully functional**  
-✅ **Permalink structure fixed**  
-✅ **Navigation working properly**  
-✅ **Admin guidance provided**  
-✅ **Diagnostic tools available**  
-
----
-
-**Catatan:** Jika masih mengalami masalah, kemungkinan ada konflik dengan plugin cache atau permalink custom. Silakan disable cache dan test kembali.
