--- wp-content/themes/mtq-aceh-pidie-jaya/GIT_COMMIT_SUMMARY.md
@@ -1,120 +1,0 @@
-# 🎉 GIT COMMIT & PUSH SUMMARY
-
-## ✅ COMMIT BERHASIL DILAKUKAN
-
-### **📊 COMMIT STATISTICS:**
-- **Commit Hash:** `bd20a46`
-- **Files Changed:** 39 files
-- **Insertions:** +656 lines
-- **Deletions:** -6,241 lines
-- **Net Reduction:** -5,585 lines (cleanup besar!)
-
-### **📤 PUSH STATUS:**
-- ✅ **Branch:** dev
-- ✅ **Remote:** origin/dev
-- ✅ **Objects:** 70 enumerated, 51 written
-- ✅ **Data Transfer:** 79.32 MiB uploaded
-- ⚠️ **Warning:** Backup file 79MB (GitHub recommends <50MB)
-
-## 📋 COMMITTED CHANGES:
-
-### **🆕 NEW FILES ADDED:**
-1. `PROTOTYPE_DELETION_SUCCESS.md` - Dokumentasi penghapusan prototype
-2. `PROTOTYPE_MIGRATION_ANALYSIS.md` - Analisis dampak migrasi
-3. `YOUTUBE_LIVE_DUPLICATE_FIX.md` - Dokumentasi perbaikan duplikasi
-4. `YOUTUBE_LIVE_STATUS.md` - Status lengkap sistem YouTube Live
-5. `assets/js/countdown.js` - JavaScript countdown (migrated)
-6. `assets/js/index.js` - JavaScript utama (migrated)
-7. `prototype-backup-20250913-112808.tar.gz` - Backup prototype folder
-
-### **📝 MODIFIED FILES:**
-1. `tailwind.config.js` - Update content paths
-2. `functions.php` - Update enqueue scripts
-3. `header.php` - Update logo path
-4. `front-page.php` - Update image paths
-5. `assets/css/youtube-live.css` - Add SVG icon support
-6. `dist/app.css` - Rebuilt CSS
-7. `inc/youtube-live-admin.php` - Fix duplicate initialization
-8. `inc/youtube-live-display.php` - Fix duplicate initialization
-9. `template-parts/youtube-live.php` - Replace Font Awesome with SVG
-
-### **🗑️ DELETED FILES:**
-- **Entire prototype/ folder** (29 files):
-  - CSS, HTML, JavaScript files
-  - Images (already migrated to assets/)
-  - Upload documents and PDFs
-  - News templates and prototypes
-
-## 🎯 MAJOR IMPROVEMENTS COMMITTED:
-
-### **1. YouTube Live System Fixes:**
-- ✅ Fixed duplicate configuration issues
-- ✅ Replaced Font Awesome with native SVG icons
-- ✅ Enhanced background color controls
-- ✅ Improved social sharing functionality
-- ✅ Added comprehensive admin panel
-
-### **2. File Structure Optimization:**
-- ✅ Migrated prototype assets to assets/ folder
-- ✅ Deleted unused prototype files
-- ✅ Organized JavaScript and CSS properly
-- ✅ Updated all asset references
-
-### **3. Performance Improvements:**
-- ✅ Removed external Font Awesome dependency
-- ✅ Cleaned up unused prototype files (-6,241 lines)
-- ✅ Optimized Tailwind configuration
-- ✅ Improved loading times
-
-### **4. Documentation & Backup:**
-- ✅ Created comprehensive documentation
-- ✅ Backup prototype folder safely
-- ✅ Detailed migration analysis
-- ✅ Step-by-step fix documentation
-
-## 📈 COMMIT HISTORY:
-
-```
-bd20a46 (HEAD -> dev) 🚀 Major Refactoring: YouTube Live System & Prototype Migration
-1436023 🎨 Remove Padding from YouTube Live Container  
-ac3acd6 🔧 Fix YouTube Live Background Settings Not Working
-8f56351 ✨ Add Background Color Controls for YouTube Live Section
-7e73efd (origin/dev) 🎨 Set YouTube Live Section Background to Transparent
-```
-
-## ⚠️ GITHUB WARNINGS:
-
-### **Large File Warning:**
-- File: `prototype-backup-20250913-112808.tar.gz` (79.27 MB)
-- GitHub Limit: 50 MB recommended
-- **Impact:** File uploaded successfully but triggers warning
-- **Recommendation:** Consider Git LFS for large files in future
-
-## 🎯 NEXT STEPS:
-
-### **Immediate Actions:**
-1. ✅ **Commit completed** - All changes safely stored
-2. ✅ **Push successful** - Changes available on remote
-3. ✅ **Backup created** - Safe rollback available
-4. ✅ **Documentation complete** - Full trail of changes
-
-### **Future Considerations:**
-1. **Testing:** Verify all functionality on live/staging site
-2. **Performance:** Monitor site performance improvements
-3. **Maintenance:** Regular cleanup of old backup files
-4. **Git LFS:** Consider for future large file storage
-
-## 🏆 SUCCESS SUMMARY:
-
-**🎉 SEMUA PERUBAHAN BERHASIL DI-COMMIT DAN PUSH!**
-
-- YouTube Live system fully optimized ✅
-- Prototype folder safely migrated & deleted ✅
-- File structure cleaned and organized ✅
-- Comprehensive documentation added ✅
-- All changes backed up and versioned ✅
-
----
-*Commit Date: September 13, 2025*  
-*Branch: dev*  
-*Status: SUCCESS*
