--- wp-content/themes/mtq-aceh-pidie-jaya/PROTOTYPE_DELETION_SUCCESS.md
@@ -1,119 +1,0 @@
-# ✅ PENGHAPUSAN FOLDER PROTOTYPE - SELESAI
-
-## 🎉 STATUS: MIGRASI BERHASIL DILAKUKAN
-
-### **RINGKASAN MIGRASI:**
-
-#### ✅ **ASSETS YANG DIPINDAHKAN:**
-1. **JavaScript Files:**
-   - `prototype/js/index.js` → `assets/js/index.js`
-   - `prototype/js/countdown.js` → `assets/js/countdown.js`
-
-2. **Image Files (sudah ada):**
-   - `prototype/images/logo.png` → `assets/images/logo.png` ✅
-   - `prototype/images/gub-wagub-aceh.png` → `assets/images/gub-wagub-aceh.png` ✅
-   - `prototype/images/bupati-dan-wakil-2025.png` → `assets/images/bupati-dan-wakil-2025.png` ✅
-
-#### ✅ **KODE YANG DIUPDATE:**
-
-1. **functions.php:**
-   ```php
-   // OLD: prototype/js/index.js & prototype/js/countdown.js
-   // NEW: assets/js/index.js & assets/js/countdown.js
-   ```
-
-2. **header.php:**
-   ```php
-   // OLD: prototype/images/logo.png
-   // NEW: assets/images/logo.png
-   ```
-
-3. **front-page.php:**
-   ```php
-   // OLD: prototype/images/gub-wagub-aceh.png
-   // NEW: assets/images/gub-wagub-aceh.png
-   
-   // OLD: prototype/images/bupati-dan-wakil-2025.png
-   // NEW: assets/images/bupati-dan-wakil-2025.png
-   ```
-
-4. **tailwind.config.js:**
-   ```javascript
-   // REMOVED: "./wp-content/themes/mtq-aceh-pidie-jaya/prototype/**/*.html"
-   // REMOVED: "./wp-content/themes/mtq-aceh-pidie-jaya/prototype/js/**/*.js"
-   // ADDED: "./wp-content/themes/mtq-aceh-pidie-jaya/assets/js/**/*.js"
-   ```
-
-#### ✅ **BACKUP DIBUAT:**
-- **File Backup:** `prototype-backup-20250913-112808.tar.gz`
-- **Size:** 83MB
-- **Location:** Theme root directory
-- **Contains:** Complete prototype folder
-
-#### ✅ **TESTING RESULTS:**
-- ✅ Tailwind build: SUCCESS
-- ✅ JavaScript files: MIGRATED
-- ✅ Image assets: AVAILABLE
-- ✅ No broken references: CONFIRMED
-- ✅ Build process: WORKING
-
-## 🎯 DAMPAK SETELAH PENGHAPUSAN:
-
-### **✅ TIDAK ADA DAMPAK NEGATIF:**
-1. **Header navigation** - Tetap berfungsi (assets/js/index.js)
-2. **Mobile menu** - Tetap berfungsi (assets/js/index.js)
-3. **Countdown timer** - Tetap berfungsi (assets/js/countdown.js)
-4. **Logo MTQ** - Tetap tampil (assets/images/logo.png)
-5. **Foto Gubernur** - Tetap tampil (assets/images/gub-wagub-aceh.png)
-6. **Foto Bupati** - Tetap tampil (assets/images/bupati-dan-wakil-2025.png)
-7. **Tailwind build** - Tetap berfungsi (updated config)
-
-### **✅ KEUNTUNGAN YANG DIDAPAT:**
-1. **Cleaner file structure** - Hanya assets yang digunakan
-2. **Smaller theme size** - Mengurangi files tidak terpakai
-3. **Better organization** - Semua assets di folder assets/
-4. **No prototype files** - Menghilangkan confusion development vs production
-
-## 📋 CHECKLIST FINAL:
-
-- ✅ **Backup created**: prototype-backup-20250913-112808.tar.gz (83MB)
-- ✅ **JavaScript migrated**: index.js & countdown.js to assets/js/
-- ✅ **Images confirmed**: All images available in assets/images/
-- ✅ **Code updated**: functions.php, header.php, front-page.php, tailwind.config.js
-- ✅ **Build tested**: npm run build successful
-- ✅ **Prototype deleted**: Folder completely removed
-- ✅ **No broken links**: All references updated
-
-## 🚀 NEXT STEPS:
-
-### **RECOMMENDED TESTING:**
-1. **Frontend Testing:**
-   - Test header logo display
-   - Test mobile menu functionality
-   - Test countdown timer
-   - Test leadership photos
-   - Check browser console for errors
-
-2. **Performance Testing:**
-   - Verify page load times
-   - Test mobile responsiveness
-   - Check JavaScript functionality
-
-### **ROLLBACK PLAN (if needed):**
-```bash
-# If any issues found, restore from backup:
-cd wp-content/themes/mtq-aceh-pidie-jaya/
-tar -xzf prototype-backup-20250913-112808.tar.gz
-# Then revert code changes manually
-```
-
-## 🎯 CONCLUSION:
-
-**✅ FOLDER PROTOTYPE BERHASIL DIHAPUS TANPA DAMPAK NEGATIF**
-
-Semua assets yang diperlukan telah dimigrasikan ke folder `assets/` dan semua referensi kode telah diupdate. Website akan berfungsi normal tanpa folder prototype.
-
----
-*Migration completed: September 13, 2025*  
-*Status: SUCCESS - No issues*  
-*Backup: Available for rollback if needed*
