--- wp-content/themes/mtq-aceh-pidie-jaya/docs/SECURITY_GUIDE.md
@@ -1,198 +1,0 @@
-# Panduan Keamanan Template MTQ Aceh Pidie Jaya
-
-## 🔒 Overview Sistem Keamanan
-
-Template ini telah dilengkapi dengan sistem keamanan berlapis yang komprehensif sesuai standar WordPress dan OWASP.
-
-### Skor Keamanan: 91/100 ⭐
-
-## 🛡️ Fitur Keamanan yang Diimplementasikan
-
-### 1. Input Sanitization & Validation
-- ✅ Semua input dari `$_POST`, `$_GET`, dan `$_REQUEST` disanitasi
-- ✅ Penggunaan `sanitize_text_field()`, `sanitize_email()`, dll
-- ✅ Validasi tipe data dan format input
-
-### 2. Output Escaping  
-- ✅ Semua output menggunakan `esc_html()`, `esc_attr()`, `esc_url()`
-- ✅ Proteksi terhadap XSS (Cross-Site Scripting)
-- ✅ Safe HTML rendering
-
-### 3. CSRF Protection
-- ✅ WordPress nonce verification di semua form
-- ✅ Nonce validation di AJAX endpoints
-- ✅ Token-based security untuk admin actions
-
-### 4. Rate Limiting
-- ✅ Pembatasan request untuk AJAX endpoints
-- ✅ IP-based rate limiting
-- ✅ Automatic blocking untuk abuse
-
-### 5. Access Control
-- ✅ Capability checking untuk admin functions
-- ✅ Role-based access control
-- ✅ Secure admin page restrictions
-
-### 6. File Security
-- ✅ `.htaccess` protection untuk direktori sensitif
-- ✅ Blocked access ke script files
-- ✅ Secure file upload handling
-
-### 7. Security Headers
-- ✅ X-Frame-Options
-- ✅ X-Content-Type-Options
-- ✅ X-XSS-Protection
-- ✅ Referrer-Policy
-
-### 8. Logging & Monitoring
-- ✅ Security event logging
-- ✅ Failed login attempt tracking
-- ✅ Suspicious activity monitoring
-
-## 📋 Checklist Keamanan untuk Developer
-
-### Input Handling
-```php
-// ✅ BENAR
-$safe_input = sanitize_text_field($_POST['user_input']);
-
-// ❌ SALAH
-$unsafe_input = $_POST['user_input'];
-```
-
-### Output Escaping
-```php
-// ✅ BENAR
-echo esc_html($user_data);
-
-// ❌ SALAH
-echo $user_data;
-```
-
-### AJAX Security
-```php
-// ✅ BENAR
-if (!wp_verify_nonce($_POST['nonce'], 'action_name')) {
-    wp_die('Security check failed');
-}
-
-// ❌ SALAH - Tidak ada nonce verification
-```
-
-## 🔧 Konfigurasi Keamanan
-
-### 1. Rate Limiting Configuration
-```php
-// Dalam functions.php
-// Maksimal 10 request per menit untuk user biasa
-// Maksimal 20 request per menit untuk admin
-```
-
-### 2. Security Headers
-Headers keamanan otomatis diterapkan untuk semua halaman:
-- `X-Frame-Options: SAMEORIGIN`
-- `X-Content-Type-Options: nosniff`
-- `X-XSS-Protection: 1; mode=block`
-
-### 3. File Protection
-Direktori berikut dilindungi dengan `.htaccess`:
-- `/scripts/` - Blokir akses langsung ke script files
-- `/data/` - Proteksi data files
-- `/logs/` - Proteksi log files
-
-## 🚨 Monitoring Keamanan
-
-### Log Security Events
-Security events dicatat di `/wp-content/debug.log`:
-```
-[2024-01-15 10:30:15] SECURITY: Rate limit exceeded - IP: 192.168.1.100
-[2024-01-15 10:32:22] SECURITY: Nonce verification failed - Action: countdown_update
-```
-
-### Dashboard Security
-Monitor aktivitas melalui:
-1. WordPress admin dashboard
-2. Security log files
-3. Rate limiting status
-
-## 🔍 Testing Keamanan
-
-### 1. XSS Testing
-```javascript
-// Test input fields dengan:
-<script>alert('XSS')</script>
-// Harus di-escape otomatis
-```
-
-### 2. CSRF Testing
-```bash
-# Test tanpa nonce - harus gagal
-curl -X POST /wp-admin/admin-ajax.php \
-  -d "action=mtq_countdown_update&data=test"
-```
-
-### 3. Rate Limit Testing
-```bash
-# Test multiple requests
-for i in {1..25}; do
-  curl -X POST /wp-admin/admin-ajax.php
-done
-# Request ke-21 dst harus diblokir
-```
-
-## 📚 Best Practices
-
-### 1. Regular Updates
-- ✅ Update WordPress core
-- ✅ Update themes & plugins
-- ✅ Monitor security patches
-
-### 2. Strong Authentication
-- ✅ Gunakan password kuat
-- ✅ Enable 2FA jika tersedia
-- ✅ Limit login attempts
-
-### 3. Backup Strategy
-- ✅ Regular database backup
-- ✅ File system backup
-- ✅ Test restore procedures
-
-### 4. Monitoring
-- ✅ Monitor error logs
-- ✅ Track suspicious activities
-- ✅ Regular security audits
-
-## 🆘 Incident Response
-
-### Jika Menemukan Masalah Keamanan:
-
-1. **Immediate Actions:**
-   - Nonaktifkan theme jika diperlukan
-   - Check security logs
-   - Identify attack vectors
-
-2. **Investigation:**
-   - Review affected files
-   - Check for malicious code
-   - Analyze access logs
-
-3. **Recovery:**
-   - Clean infected files
-   - Update security measures
-   - Strengthen affected components
-
-4. **Prevention:**
-   - Patch vulnerabilities
-   - Update security rules
-   - Enhance monitoring
-
-## 📞 Support
-
-Untuk pertanyaan keamanan atau laporan bug:
-- Email: security@pidiejayakab.go.id
-- Priority: HIGH untuk security issues
-- Response time: < 24 jam
-
----
-*Last updated: January 2024*
-*Security Audit Score: 91/100*
